temporarily consolidate all in one file
it is too hard to swap between the choc file and the keyboard.
So for now putting all into the same file.

After this is complete, will convert back to an additive model and
just have a single key. (that way it is easier to have non grid
and non-uniform patterns